requirements.txt
streamlit
transformers
torch
pyttsx3
requests
huggingface_hub
openai>=1.0.0
python-dotenv
